requirements.dev.txt
flake8>=3.9.2,<3.10
djangorestframework-types>=0.9.0